Deprecate the glob fields in the DICOM converter
Now, the DICOM files in the RAWDATA folder are discovered
automatically. A function get_dicomfiles(modality) returns
the list of discovered DICOM files, the first one is used
as an input to diff_unpack.

The extraction of metadata from the DICOM files like bvalues
and bvectors is disabled, because there is no generic way
to retrieve this information from DICOM files from different
sources.